// src/modules/survey/statistic/types/SurveyStatisticTypes.tsx
export interface Selection {
  surveyPostAt: string;
  userNickname: string;
  surveyNo: number;
  surveyTitle: string;
  surveyQuestionNo: number;
  surveyQuestionTitle: string;
  questionTypeNo: number;
  selectionNo: number;
  selectionValue: string;
  selectionCount: number;
  surveySubjectiveAnswer: string;
  surveySubjectiveAnswerCount: number;
  surveyWriter: string;
}